+++ src/FsCodec.Box/CoreCodec.fs
@@ -22,7 +22,7 @@ type Codec private () =
             // <summary>Maps a fresh Event resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c><br/>
             // The function is also expected to derive an optional <c>meta</c> object that will be serialized with the same <c>encoder</c>,
             // and <c>eventId</c>, <c>correlationId</c>, <c>causationId</c> and an Event Creation<c>timestamp</c></summary>.
-            down : struct ('Context * 'Event) -> struct ('Contract * 'Meta voption * Guid * string * string * DateTimeOffset voption),
+            down : struct ('Context * 'Event) -> struct ('Contract * 'Meta voption * Guid * string * string * DateTimeOffset),
             // <summary>Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them.</summary>
             [<Optional; DefaultParameterValue(null)>] ?rejectNullaryCases)
         : FsCodec.IEventCodec<'Event, 'Body, 'Context> =
@@ -36,12 +36,12 @@ type Codec private () =
                 allowNullaryCases = not (defaultArg rejectNullaryCases false))
 
         { new FsCodec.IEventCodec<'Event, 'Body, 'Context> with
+
             member _.Encode(context, event) =
-                let struct (c, meta : 'Meta voption, eventId, correlationId, causationId, timestamp : DateTimeOffset voption) = down struct (context, event)
+                let struct (c, meta : 'Meta voption, eventId, correlationId, causationId, timestamp) = down (context, event)
                 let enc = dataCodec.Encode c
-                let meta' = match meta with ValueSome x -> encoder.Encode<'Meta> x | ValueNone -> Unchecked.defaultof<_>
-                let ts = match timestamp with ValueNone -> None | ValueSome v -> Some v
-                EventData.Create(enc.CaseName, enc.Payload, meta', eventId, correlationId, causationId, ?timestamp = ts)
+                let meta' = match meta with ValueSome x -> encoder.Encode<'Meta> x | ValueNone -> Unchecked.defaultof<'Body>
+                EventData(enc.CaseName, enc.Payload, meta', eventId, correlationId, causationId, timestamp)
 
             member _.TryDecode encoded =
                 match dataCodec.TryDecode { CaseName = encoded.EventType; Payload = encoded.Data } with
@@ -61,9 +61,9 @@ type Codec private () =
             // <summary>Maps a fresh Event resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
             // The function is also expected to derive
             //   a <c>meta</c> object that will be serialized with the same options (if it's not <c>None</c>)
-            //   and an Event Creation <c>timestamp</c>.</summary>
+            //   and an Event Creation <c>timestamp</c>  (Default: DateTimeOffset.UtcNow).</summary>
             down : 'Event -> struct ('Contract * 'Meta voption * DateTimeOffset voption),
-            // <summary>Uses the 'Context passed to the Encode call and the 'Meta emitted by <c>down</c> to a) the final metadata b) the <c>eventId</c> c) the <c>correlationId</c> and d) the <c>causationId</c></summary>
+            // <summary>Uses the 'Context passed to the Encode call and the 'Meta emitted by <c>down</c> to produce a) the final metadata b) the <c>eventId</c> c) the <c>correlationId</c> and d) the <c>causationId</c></summary>
             mapCausation : struct ('Context * 'Meta voption) -> struct ('Meta voption * Guid * string * string),
             // <summary>Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them.</summary>
             [<Optional; DefaultParameterValue(null)>] ?rejectNullaryCases)
@@ -72,7 +72,7 @@ type Codec private () =
         let down struct (context, union) =
             let struct (c, m, t) = down union
             let struct (m', eventId, correlationId, causationId) = mapCausation (context, m)
-            struct (c, m', eventId, correlationId, causationId, t)
+            struct (c, m', eventId, correlationId, causationId, match t with ValueSome t -> t | ValueNone -> DateTimeOffset.Now)
         Codec.Create(encoder, up = up, down = down, ?rejectNullaryCases = rejectNullaryCases)
 
     /// <summary>Generate an <code>IEventCodec</code> using the supplied <c>encoder</c>.<br/>
@@ -88,7 +88,7 @@ type Codec private () =
             // <summary>Maps a fresh <c>'Event</c> resulting from a Decision in the Domain representation type down to the TypeShape <c>UnionConverter</c> <c>'Contract</c>
             // The function is also expected to derive
             //   a <c>meta</c> object that will be serialized with the same options (if it's not <c>None</c>)
-            //   and an Event Creation <c>timestamp</c>.</summary>
+            //   and an Event Creation <c>timestamp</c> (Default: DateTimeOffset.UtcNow).</summary>
             down : 'Event -> struct ('Contract * 'Meta voption * DateTimeOffset voption),
             // <summary>Enables one to fail encoder generation if union contains nullary cases. Defaults to <c>false</c>, i.e. permitting them.</summary>
             [<Optional; DefaultParameterValue(null)>] ?rejectNullaryCases)
@@ -107,5 +107,5 @@ type Codec private () =
         : FsCodec.IEventCodec<'Union, 'Body, unit> =
 
         let up struct (_e : FsCodec.ITimelineEvent<'Body>, u : 'Union) : 'Union = u
-        let down (event : 'Union) = struct (event, ValueNone, ValueNone)
+        let down (event : 'Union) = struct (event, ValueNone (*Meta*), ValueNone (*Timestamp*))
         Codec.Create(encoder, up = up, down = down, ?rejectNullaryCases = rejectNullaryCases)