+++ src/FsCodec/FsCodec.fs
@@ -44,11 +44,11 @@ open System
 
 /// An Event about to be written, see <c>IEventData<c> for further information
 [<NoComparison; NoEquality>]
-type EventData<'Format> private (eventType, data, meta, eventId, correlationId, causationId, timestamp) =
+type EventData<'Format>(eventType, data, meta, eventId, correlationId, causationId, timestamp) =
 
     static member Create(eventType, data, ?meta, ?eventId, ?correlationId, ?causationId, ?timestamp : DateTimeOffset) : IEventData<'Format> =
-        let meta =    match meta      with Some x -> x   | None -> Unchecked.defaultof<_>
-        let eventId = match eventId   with Some x -> x   | None -> Guid.Empty
+        let meta =    match meta      with Some x -> x   | None -> Unchecked.defaultof<'Format>
+        let eventId = match eventId   with Some x -> x   | None -> Guid.NewGuid()
         let ts =      match timestamp with Some ts -> ts | None -> DateTimeOffset.UtcNow
         EventData(eventType, data, meta, eventId, Option.toObj correlationId, Option.toObj causationId, ts) :> _
 
@@ -74,7 +74,7 @@ type EventData<'Format> private (eventType, data, meta, eventId, correlationId, 
 
 /// An Event or Unfold that's been read from a Store and hence has a defined <c>Index</c> on the Event Timeline
 [<NoComparison; NoEquality>]
-type TimelineEvent<'Format> private (index, eventType, data, meta, eventId, correlationId, causationId, timestamp, isUnfold, context, size) =
+type TimelineEvent<'Format>(index, eventType, data, meta, eventId, correlationId, causationId, timestamp, isUnfold, context, size) =
 
     static member Create(index, eventType, data, ?meta, ?eventId, ?correlationId, ?causationId, ?timestamp, ?isUnfold, ?context, ?size) : ITimelineEvent<'Format> =
         let isUnfold = defaultArg isUnfold false
@@ -122,6 +122,7 @@ type EventCodec<'Event, 'Format, 'Context> private () =
         let downConvert = TimelineEvent.Map down
 
         { new IEventCodec<'Event, 'TargetFormat, 'Context> with
+
             member _.Encode(context, event) =
                 let encoded = native.Encode(context, event)
                 upConvert encoded